requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.1
billiard==4.2.0
boto3==1.35.63
botocore==1.35.63
cachetools==5.5.0
celery==5.4.0
certifi==2023.7.22
cffi==1.17.1
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.1
Django==4.2.4
django-admin-list-filter-dropdown==1.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-cleanup==8.1.0
django-colorfield==0.9.0
django-cors-headers==4.2.0
django-countries==7.6.1
django-filter==25.1
django-formtools==2.5.1
django-js-asset==2.1.0
django-phonenumber-field==8.0.0
django-ratelimit==4.1.0
django-ses==4.2.0
django-ssl-smtp==2.0.0
django-storages==1.14.6
django-timezone-field==7.0
django-two-factor-auth==1.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
google-auth==2.34.0
gunicorn==23.0.0
h11==0.16.0
honcho==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
jmespath==1.0.1
kombu==5.4.0
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.6
packaging==25.0
paypalrestsdk==1.13.3
pexpect==4.9.0
phonenumbers==8.13.45
pillow==11.2.1
pillow_heif==0.22.0
prompt_toolkit==3.0.47
protobuf==4.21.12
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMuPDF==1.23.17
PyMuPDFb==1.23.9
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3
qrcode==7.4.2
redis==5.0.8
requests==2.31.0
rsa==4.9
s3transfer==0.10.3
service-identity==24.2.0
six==1.16.0
sqlparse==0.4.4
stripe==12.1.0
supervisor==4.2.5
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.0.4
uvicorn[standard]==0.34.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.9.4
zope.interface==7.2
django-otp==1.6.0
django-extensions==3.2.3
django-cryptography>=1.1
openai>=1.12.0
httpx==0.27.2
reportlab==4.2.2
openpyxl==3.1.5